requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-autoslug==1.9.8
django-countries==7.2.1
django-js-asset==1.2.2
django-mptt==0.13.2
django-session-timeout==0.1.0
idna==3.2
Pillow==8.3.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
